+++ src/main/java/com/github/idea/ginkgo/GinkgoSpecType.java
@@ -2,6 +2,8 @@ package com.github.idea.ginkgo;
 
 import com.github.idea.ginkgo.util.GinkgoUtil;
 
+import java.util.EnumSet;
+
 public enum GinkgoSpecType {
     DESCRIBE("Describe"),
     DESCRIBE_TABLE("DescribeTable"),
@@ -30,5 +32,9 @@ public enum GinkgoSpecType {
 
     public GinkgoPendingSpecType getPendingSpecType() {
         return GinkgoUtil.getGinkgoPendingSpecType("P"+specType);
+    }
+
+    public static boolean isGinkgoSpec(String spec) {
+        return EnumSet.allOf(GinkgoSpecType.class).stream().anyMatch(e -> e.specType().equals(spec));
     }
 }